requirements.txt
pandas
kaggle
kagglehub
python-dotenv
psycopg2-binary
SQLAlchemy
dbt-core
dbt-postgres
